commit for febonaccis2

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -5,20 +5,20 @@
  */
 int main(void)
 {
-	int fp1;
-	int fp2;
-	int fn;
-	int i;
+	long fp1;
+	long fp2;
+	long fn;
+	long i;
 
 	fp1 = 1;
 	fp2 = 2;
 	i = 3;
-	printf("%d%c%c", fp1, 44, 32);
-	printf("%d%c%c", fp2, 44, 32);
+	printf("%ld%c%c", fp1, 44, 32);
+	printf("%ld%c%c", fp2, 44, 32);
 	while (i <= 50)
 	{
 		fn = fp1 + fp2;
-		printf("%d%c%c", fn, 44, 32);
+		printf("%ld%c%c", fn, 44, 32);
 		fp1 = fp2;
 		fp2 = fn;
 		i++;
